Refactorización de importaciones: se actualiza el uso del hook useEscuela a useEscuelaStore en las páginas de calendario y ciclos escolares.
